% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thePlot.R
\name{thePlot}
\alias{thePlot}
\title{thePlot}
\usage{
thePlot(data, x, y = NULL, group = NULL)
}
\arguments{
\item{data}{The dataset from where the variables will be taken from}

\item{x}{The first variable}

\item{y}{The second variable}

\item{group}{The variable to group the plot with}
}
\value{
A plot
}
\description{
This \code{thePlot} takes nearly all variables and returns you a plot
}
\examples{
\dontrun{
library(ggplot2)
data(mtcars)
thePlot(mtcars, hp)
}
}
