% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacPlot-package.R
\docType{package}
\name{qacPlot-package}
\alias{qacPlot}
\alias{qacPlot-package}
\title{qacPlot: Takes in variables, determines what class they are, and create a simple plot}
\description{
This package will take in three variables with the third variable as the grouping variable. The package will then determine whether they are a categorical or numerical variable and make appropriate plots.
}
\author{
\strong{Maintainer}: Calvin Gao \email{cgao@wesleyan.edu}

Authors:
\itemize{
  \item Terry Yuan \email{yuan@wesleyan.edu}
}

}
\keyword{internal}
